Refactor keyboard input
